# src/docs/swagger.yaml
basePath: /api/magic-circle
definitions:
  auth.CreateAdminReq:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  auth.CreateAdminResp:
    properties:
      id:
        type: integer
      intstituteId:
        type: integer
      login:
        type: string
    type: object
  auth.LoginReq:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  auth.LoginResp:
    properties:
      expire:
        type: string
      token:
        type: string
    type: object
  direction.Direction:
    properties:
      id:
        type: integer
      name:
        type: string
      profiles:
        items:
          $ref: '#/definitions/profile.Profile'
        type: array
    type: object
...
  institute.GetInstitutesResp:
    properties:
      institutes:
        items:
          $ref: '#/definitions/institute.Institute'
        type: array
    type: object
  institute.Institute:
    properties:
      directions:
        items:
          $ref: '#/definitions/direction.Direction'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  institute.UpdateInstituteReq:
    properties:
      name:
        type: string
    type: object
  institute.UpdateInstituteResp:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  profile.Profile:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a server to get projects from github
  title: Magic-Circle API
  version: "1.0"
paths:
...
  /v1/auth/admin/{intstitute_id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.CreateAdminReq'
      - description: id of institute
        in: path
        name: intstitute_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.CreateAdminResp'
      security:
      - ApiKeyAuth: []
      summary: create admin
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResp'
      summary: login admin or super admin
      tags:
      - auth
  /v1/auth/refreshToken:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResp'
      security:
      - ApiKeyAuth: []
      summary: login admin or super admin
      tags:
      - auth
...
  /v1/institute:
    get:
      description: return all institutes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/institute.GetInstitutesResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all institutes
      tags:
      - institute
  /v1/institute/{id}:
    delete:
      description: Delete Institute by id
      parameters:
      - description: id of institute
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Institute by id
      tags:
      - institute
    put:
      consumes:
      - application/json
      description: update institute
      parameters:
      - description: id of institute
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/institute.UpdateInstituteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/institute.UpdateInstituteResp'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Update institute
      tags:
      - institute
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"